+++ manager-api/src/api/accounts/accounts.module.ts
@@ -1,6 +1,7 @@
 import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { AclModule } from "../../core/acl/acl.module";
+import { JwtAuthModule } from "../../core/auth/jwt/jwt.module";
 import { CustomPermissionGuardModule } from "../../core/custom-permission-guard/custom-permission-guard.module";
 import { AccountInvitation } from "../../core/entities/account-invitation.entity";
 import { Account } from "../../core/entities/account.entity";
@@ -21,6 +22,7 @@ import { AccountsService } from "./accounts.service";
     CustomPermissionGuardModule,
     AclModule,
     GeocodingModule,
+    JwtAuthModule,
   ],
   providers: [AccountsService],
   controllers: [AccountsController],